Crushroom and old Magma crawler
adding back old magma crawler for now until new one is properly worked
out

+++ src/main/java/erebus/entity/EntityCentipede.java
@@ -170,7 +170,7 @@ public class EntityCentipede extends EntityMob {
 	}
 
 	public int getSkin() {
-		return dataManager.get(SKIN_TYPE).intValue();
+		return dataManager.get(SKIN_TYPE);
 	}
 
 	@Override
